fix(test): remove test_driver_get_or_create
